graphrag/memgraph_assistant/backend/pyproject.toml
[project]
name = "memgraph-assistant-backend"
version = "0.1.0"
description = "Memgraph Assistant Backend - Documentation-aware chat assistant"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.109.0",
    "httpx>=0.28.1",
    "lightrag-hku[api]==1.4.8.2",
    "litellm>=1.77.1",
    "mcp>=1.14.1",
    "neo4j>=5.28.1",
    "numpy>=2.2.6",
    "openai-agents==0.6.1",
    "python-dotenv>=1.1.1",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "unstructured[all-docs]>=0.12.6",
    "uvicorn[standard]>=0.31.1",
]

